change version dependency and optimize create

+++ crates/iv_domain/src/user/response.rs
@@ -2,6 +2,6 @@ use serde::Serialize;
 
 #[derive(Serialize)]
 pub struct ResponseCreateUser {
-  pub id: i32,
+  pub id: i64,
   pub username: String,
 }
